backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "backend for icb score app"
authors = ["Roddy Brown <roddy@notroddy.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
django = "^5.0.2"
psycopg = "^3.1.18"
gunicorn = "^21.2.0"
pytesseract = "^0.3.10"
numpy = "^2.0.0"
opencv-python = "^4.10.0.84"
openai = "^1.35.10"
pandas = "^2.2.2"
questionary = "^2.0.1"
psycopg2-binary = "^2.9.9"
django-rest-framework = "^0.1.0"
django-extensions = "^3.2.3"
ipython = "^8.31.0"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.6.0"
python-dotenv = "^1.0.1"
django-db-utilities = "0.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"